// app/utils/questionType.enum.ts
export enum QUESTION_TYPE {
    SHORT_ANSWER = "663f0b3baa8c637f5b0d37d0",
    PARAGRAPH = "663f0b4eaa8c637f5b0d37d2",
    MULTIPLE_CHOICE = "663f0b5caa8c637f5b0d37d4",
    CHECKBOX = "663f0b6daa8c637f5b0d37d6",
    DROPDOWN = "663f0b77aa8c637f5b0d37d8",
    LINEAR_SCALE = "663f0b89aa8c637f5b0d37da",
    MULTIPLE_CHOICE_GRID = "663f0b98aa8c637f5b0d37dc",
    CHECKBOX_GRID = "663f0ba4aa8c637f5b0d37de",
    DATE = "663f0bb1aa8c637f5b0d37e0",
    TIME = "663f0bb9aa8c637f5b0d37e2"
}